Refactor Inventory Plugin module to import from module_utils
